+++ src/aswwu/alchemy_new/elections.py
@@ -5,8 +5,8 @@ import logging
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-import aswwu.models.bases as base
-import aswwu.models.elections as elections_model
+import src.aswwu.models.bases as base
+import src.aswwu.models.elections as elections_model
 from datetime import datetime
 from settings import database
