+++ src/aswwu/models/elections.py
@@ -2,8 +2,8 @@ from datetime import datetime
 
 from sqlalchemy import Column, ForeignKey, String, DateTime, Boolean, Integer
 
-from aswwu.models.bases import ElectionBase
-import aswwu.alchemy_new.mask as mask_alchemy
+from src.aswwu.models.bases import ElectionBase
+import src.aswwu.alchemy_new.mask as mask_alchemy
 
 
 class Election(ElectionBase):
