+++ src/aswwu/route_handlers/pages.py
@@ -1,13 +1,12 @@
 import logging
 
-import bleach
 import datetime
 import tornado.web
 import json
 
-import aswwu.alchemy_new.pages as alchemy
-import aswwu.models.pages as pages_model
-from aswwu.base_handlers import BaseHandler
+import src.aswwu.alchemy_new.pages as alchemy
+import src.aswwu.models.pages as pages_model
+from src.aswwu.base_handlers import BaseHandler
 
 logger = logging.getLogger("aswwu")
