+++ src/aswwu/validators/elections.py
@@ -3,9 +3,9 @@
 
 from datetime import datetime
 
-import aswwu.alchemy_new.elections as elections_alchemy
-import aswwu.alchemy_new.mask as mask_alchemy
-import aswwu.exceptions as exceptions
+from src import aswwu as elections_alchemy
+import src.aswwu.alchemy_new.mask as mask_alchemy
+import src.aswwu.exceptions as exceptions
 
 datetime_format = '%Y-%m-%d %H:%M:%S.%f'
